emdad-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.6.0
pyjwt==2.9.0
pydantic==2.9.2
python-multipart==0.0.9
